Update README.md
This Note keeping app will allow the user to record their thoughts and ideas. The user will be able to save and delete notes and personalise the app by switching between light and dark colour themes and adding dividing lines between the note previews on the home screen. In creating this app, you will learn many of the fundamentals of Android programming using Kotlin and XML.